fix: relative paths
If a relative path pointed to a module in the root of the working directory, it was possible for a global module to take precedence (if already loaded).

For example, require('../fiber') from @./app/init.lua would return package.loaded['fiber'] instead of @./fiber.lua (which may not be intended if a global module named 'fiber' was loaded earlier).

To fix that, relative modules retain their leading ./ within the `package.loaded` table.

Also, ~/ can be used to resolve a path relative to the root of the working directory. So in our example, we could use require('~/fiber') instead.