fix: init entry module
To fix relative `require` calls in the entry module, its file is put in the working directory (instead of a temporary directory). The file is named "__entry__.lua".

Additionally, we switch to the `root` before creating the entry module, so all `inject` paths are relative to `root` if specified.

Finally, always generate an entry module even when `inject` is undefined. This means the relative path loader is always available.